Addressed the comments Add BaseAccountCommand abstract class
